Add user first and last name to API and UI
In the ApiService and ApiUsersController, two properties - first name and last name - were added to the User objects. The TokenDto was updated to include this new data. Consequently, changes were also made in MainPage.xaml and MainPage.xaml.cs to display the full name of the user. A new directory called 'Api' was also added to the project structure.