Update UI elements and translations in various pages
This commit restructures the formatting in the LoginPage.xaml for better readability. It also updates translations for the titles in the AppShell.xaml. Additionally, slight coding adjustments are made in MainPage.xaml.cs while more considerable UI enhancements are implemented in MainPage.xaml and ProfilePage.xaml, for a richer user experience.