[web] Implement GetWorkouts functionality in the WorkoutController
The WorkoutController now uses the GetWorkoutsQuery to retrieve workouts for a particular user instead of using a hardcoded string. Also, a new WorkoutDto, which is used to transfer data about workouts, and a corresponding query handler GetWorkoutsQueryHandler have been added. Corresponding view and model changes have been made to support these modifications. In addition, a new WeekDays enum was introduced.