[web] Add user-specific functionality to ExerciseDetails
This update adds a user context to the ExerciseDetails view by displaying whether the currently logged-in user has selected the exercise. The code to handle the user-specific exercise selection was added to the ExerciseApiController. Consequently, adjustments were made to the GetExerciseDetails query classes to include active user ID and the status of the exercise (whether selected or not) for the active user.