[web] Add workout management functionality in calendar
Two new commands and their handlers have been added to manage workouts in the calendar: one to set a selected workout as done, another to delete a selected workout. The workflow in UI was subsequently updated to support these changes. This also involved updating the WorkoutCalendarViewModel to include a list of Workouts and a WorkoutCalendar object.