Remove JWT authentication related files and dependencies
This commit removes all the files and dependencies related to JWT token generation and user authentication process. This includes the deletion of several classes like 'AuthenticateRequest', 'AuthenticateResponse', as well as the removal of 'IJwtService' interface and its implementation. Additionally, this also involves removing the usage of these files in other parts of the code.